Migrate Report>Scans to Assets>Search.
