+++ nector_home/templates/nector_home/assets-tools.html
@@ -1,4 +1,4 @@
 <p><a href="/hosts/">Br0wSE</a></p>
 <p><a href="/hosts/ports/">P0rt5</a></p>
 <p><a href="/hosts/alerts/">A13rTs</a></p>
-<p><a href="/scans/">S34rCh</a></p>
+<p><a href="/hosts/custom-search/">S34rCh</a></p>
