+++ scans/urls.py
@@ -3,5 +3,4 @@ from . import views
 
 urlpatterns = [
     url(r'^$', views.index, name='index'),
-    url(r'^export/$', views.export, name='export'),
 ]
